Move update terrain and update position logic
On Tile.gd. Moved the update_terrain and update_position logic from
init() to _ready(). This was done in order to include setting terrain
texture updates in the latter which was not done in the first
implementation. Doing this allows us to call the _update_terrain
function when the player clicks the tile in TerraForm mode.